Created new-comp component and downsized new-group
