+++ src/views/dashboard/sections/dashboard-new-comp/DashboardNewCompStyles.tsx
@@ -1,14 +1,14 @@
 import styled from "styled-components";
 import { device } from "../../../../mediaQueries";
 
-const DashboardNewGroupWrapper = styled.div`
+const DashboardNewCompWrapper = styled.div`
   display: flex;
   justify-content: center;
-  flex-wrap: wrap;
 
   @media ${device.laptop} {
     width: 1000px;
     margin: 0 auto;
+    flex-wrap: wrap;
   }
 
   .button {
@@ -21,11 +21,9 @@ const DashboardNewGroupWrapper = styled.div`
     }
   }
   .title {
-    display: block;
-    width: 100%;
-    margin-top: 100px;
-
+    display: none;
     @media ${device.laptop} {
+      display: block;
       width: 100%;
       margin-top: 100px;
     }
@@ -54,4 +52,4 @@ const Form = styled.form`
   }
 `;
 
-export { DashboardNewGroupWrapper, Form };
+export { DashboardNewCompWrapper, Form };
