+++ src/views/dashboard/sections/dashboard-new-comp/newCompReducer.ts
@@ -6,9 +6,13 @@ interface Action {
 }
 interface State {}
 const initialState = {
-  groupName: "",
+  messageGroup: "",
+  betAmount: "",
+  startDate: moment().format("YYYY-MM-DD"),
+  endDate: moment().format("YYYY-MM-DD"),
+  goal: "",
 };
-const newGroupReducer = (state: State, action: Action) => {
+const newCompReducer = (state: State, action: Action) => {
   switch (action.type) {
     case "SET_TEXT": {
       return {
@@ -16,9 +20,15 @@ const newGroupReducer = (state: State, action: Action) => {
         [action.inputName]: action.value,
       };
     }
+    case "SET_CURRENCY": {
+      return {
+        ...state,
+        [action.inputName]: action.value,
+      };
+    }
     default: {
       throw Error(
-        `newGroupReducer got an unknown action. Error: ${JSON.stringify(
+        `newCompReducer got an unknown action. Error: ${JSON.stringify(
           action,
           undefined,
         )}`,
@@ -30,6 +40,6 @@ const newGroupReducer = (state: State, action: Action) => {
 export default useReducer.bind(
   //@ts-ignore
   this,
-  newGroupReducer,
+  newCompReducer,
   initialState,
 );
